simplify welcomepage + move util + add index files
