+++ src/data/db.json
@@ -1,4 +1,4 @@
 {
-    "notes": [],
-    "classes": []
+  "notes": [],
+  "classes": []
 }
