+++ src/pages/Error/index.ts
@@ -1,0 +1,1 @@
+export { default as ErrorPage } from "./Error";
