--- src/pages/Welcome/components/RegisterForm/styles.module.scss
@@ -1,27 +1,0 @@
-@import "../../../../assets/variables";
-
-.placeholder {
-    padding: 2em 0 1em 0;
-    font-size: 2rem;
-    text-align: center;
-}
-.form {
-    display: flex;
-    justify-content: center;
-}
-
-@media only screen and (max-width: 370px) {
-    form {
-        margin: 0 2em;
-    }
-}
-
-.register_input {
-    height: 3rem;
-    border-radius: 0.5em 0 0 0.5em !important;
-    max-width: 10em;
-}
-
-.button {
-    height: 3rem;
-}
