+++ src/pages/Welcome/index.ts
@@ -1,0 +1,1 @@
+export { default as WelcomePage } from "./Welcome";
